small cleanup of PatternRecognition and producer, for llvm compile errors
